# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: HydroGuard water pipeline monitoring and demonstration API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alerts/{id}/acknowledge:
    post:
      operationId: acknowledgeAlert
      tags: [alerts]
      summary: Acknowledge an alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Alert not found
  /alerts/{id}/assign:
    post:
      operationId: assignAlert
      tags: [alerts]
      summary: Assign an alert to a Water Champion
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertAssignmentInput"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Alert or Water Champion not found
  /alerts/{id}/resolve:
    post:
      operationId: resolveAlert
      tags: [alerts]
      summary: Resolve an active alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resolved alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Alert not found
  /alerts/{id}/maintenance:
    post:
      operationId: startAlertMaintenance
      tags: [alerts]
      summary: Create a repair from an alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertMaintenanceInput"
      responses:
        "201":
          description: Maintenance record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRecord"
        "404":
          description: Alert not found
...
  /analytics:
    get:
      operationId: getAnalytics
      tags: [monitoring]
      summary: Get chart series for an interval
      parameters:
        - in: query
          name: range
          required: false
          schema:
            type: string
            enum: [24h, 7d, 30d]
            default: 24h
      responses:
        "200":
          description: Time series and system health metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analytics"
  /reports:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report rows for export
      parameters:
        - in: query
          name: type
          required: true
          schema:
            type: string
            enum: [water-usage, water-loss, leak-events, pipeline-health, sensor-health, maintenance]
        - in: query
          name: from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: to
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Report rows and metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportData"
...
components:
  schemas:
...
    Alert:
      type: object
      required: [id, zone, location, sensorId, type, severity, flowLpm, pressureBar, createdAt, status, championName, message]
      properties:
        id: { type: string }
        zone: { type: string }
        location: { type: string }
        sensorId: { type: string }
        type: { type: string }
        severity: { type: string, enum: [critical, warning, resolved] }
        flowLpm: { type: number }
        pressureBar: { type: number }
        createdAt: { type: string, format: date-time }
        status: { type: string, enum: [active, acknowledged, assigned, in_progress, resolved] }
        championName: { type: string }
        message: { type: string }
        championId: { type: ["string", "null"] }
        riskLevel: { type: ["string", "null"], enum: [low, medium, high, null] }
        riskTitle: { type: ["string", "null"] }
        source: { type: string, enum: [field, simulator] }
        simulated: { type: boolean }
        simulatorId: { type: ["string", "null"] }
        mockSmsMessage: { type: ["string", "null"] }
...
    AlertAssignmentInput:
      type: object
      required: [championId]
      properties:
        championId: { type: string }
    AlertMaintenanceInput:
      type: object
      properties:
        championId: { type: ["string", "null"] }
        priority: { type: string, enum: [low, medium, high, critical] }
...
    MaintenanceRecord:
      type: object
      required: [id, maintenanceId, pipeline, location, issue, championName, priority, startedAt, completedAt, status]
      properties:
        id: { type: string }
        maintenanceId: { type: string }
        pipeline: { type: string }
        location: { type: string }
        issue: { type: string }
        championId: { type: ["string", "null"] }
        championName: { type: string }
        priority: { type: string, enum: [low, medium, high, critical] }
        startedAt: { type: string, format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        status: { type: string, enum: [pending, assigned, in-progress, completed] }
        alertId: { type: ["string", "null"] }
...
    Analytics:
      type: object
      required: [range, kpis, flowPressure, tankLevel, alertsByRisk, fieldResponse, alertsByZone, simulatedWaterLostPerZone]
      properties:
        range: { type: string, enum: [24h, 7d, 30d] }
        kpis:
          type: object
          required: [leakEvents, meanResponseMinutes, sensorsReporting, totalSensors]
          properties:
            leakEvents: { type: integer }
            meanResponseMinutes: { type: ["number", "string"] }
            sensorsReporting: { type: integer }
            totalSensors: { type: integer }
        flowPressure:
          type: array
          items:
            $ref: "#/components/schemas/FlowPressureZonePoint"
        tankLevel:
          type: array
          items:
            $ref: "#/components/schemas/TankLevelPoint"
        alertsByRisk:
          type: array
          items:
            $ref: "#/components/schemas/AlertRiskPoint"
        fieldResponse:
          type: array
          items:
            $ref: "#/components/schemas/FieldResponsePoint"
        alertsByZone:
          type: array
          items:
            $ref: "#/components/schemas/AlertZonePoint"
        simulatedWaterLostPerZone:
          type: array
          items:
            $ref: "#/components/schemas/SimulatedLossPoint"
    FlowPressureZonePoint:
      type: object
      required: [time, zone, flowLpm, pressureBar]
      properties:
        time: { type: string }
        zone: { type: string }
        flowLpm: { type: number }
        pressureBar: { type: number }
    TankLevelPoint:
      type: object
      required: [time, tankName, tankLevelPercent]
      properties:
        time: { type: string }
        tankName: { type: string }
        tankLevelPercent: { type: number }
    AlertRiskPoint:
      type: object
      required: [time, riskLevel, source, count]
      properties:
        time: { type: string }
        riskLevel: { type: string, enum: [low, medium, high] }
        source: { type: string, enum: [field, simulated] }
        count: { type: integer }
    FieldResponsePoint:
      type: object
      required: [championName, averageResponseMinutes]
      properties:
        championName: { type: string }
        averageResponseMinutes: { type: number }
    AlertZonePoint:
      type: object
      required: [zone, count]
      properties:
        zone: { type: string }
        count: { type: integer }
    SimulatedLossPoint:
      type: object
      required: [zone, litresLost, label]
      properties:
        zone: { type: string }
        litresLost: { type: number }
        label: { type: string }
    ReportData:
      type: object
      required: [title, from, to, generatedAt, columns, rows]
      properties:
        title: { type: string }
        from: { type: string, format: date }
        to: { type: string, format: date }
        generatedAt: { type: string, format: date-time }
        columns:
          type: array
          items: { type: string }
        rows:
          type: array
          items:
            type: array
            items: { type: string }
...