# lib/api-zod/src/generated/types/alertZonePoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HydroGuard water pipeline monitoring and demonstration API
 * OpenAPI spec version: 0.2.0
 */

export interface AlertZonePoint {
  zone: string;
  count: number;
}