# lib/api-zod/src/generated/types/flowPressureZonePoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HydroGuard water pipeline monitoring and demonstration API
 * OpenAPI spec version: 0.2.0
 */

export interface FlowPressureZonePoint {
  time: string;
  zone: string;
  flowLpm: number;
  pressureBar: number;
}