# lib/api-zod/src/generated/types/reportData.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * HydroGuard water pipeline monitoring and demonstration API
 * OpenAPI spec version: 0.2.0
 */

export interface ReportData {
  title: string;
  from: Date;
  to: Date;
  generatedAt: Date;
  columns: string[];
  rows: string[][];
}